refactor: 헤더 컴포넌트 분리 및 페이지네이션 경로 수정, 스타일 추가 (#36)
